extract CommonServerGetSideUserProp, add getServerSideProps for all pages
