+++ util/user.ts
@@ -1,6 +1,5 @@
 import {RoutePath, UserLocalStorageKey} from "./const";
-import {GetServerSidePropsContext, PreviewData} from "next";
-import {ParsedUrlQuery} from "querystring";
+import {GetServerSideProps,} from "next";
 import {getUserInfo} from "../api/user";
 
 export interface User {
@@ -33,4 +32,42 @@ export function GetLocalUserInfo(): User | null {
     return JSON.parse(userInfo)
   }
   return null
+}
+
+export const CommonServerGetSideUserProp = (toHomePage: boolean): GetServerSideProps => {
+  return async (context) => {
+    return await getUserInfo(context.req.headers.cookie).then((resp) => {
+      if (resp.headers["set-cookie"]) {
+        context.res.setHeader('set-cookie', resp.headers["set-cookie"])
+      }
+      if (resp.data.data.user.user_register_type == "email" && !resp.data.data.user.email_active) {
+        return {
+          redirect: {
+            destination: RoutePath.EmailActivateSendPage,
+            permanent: false,
+          }
+        }
+      }
+      if (toHomePage) {
+        return {
+          redirect: {
+            destination: `${RoutePath.HomePage}?page=1`,
+            permanent: false,
+          }
+        }
+      }
+      return {
+        props: {
+          user: resp.data.data.user
+        }
+      }
+    }).catch(() => {
+      return {
+        redirect: {
+          destination: RoutePath.LoginPage,
+          permanent: false,
+        }
+      }
+    })
+  }
 }
